Display directories before files, then sort by alphabetical order
